Move country label handling to separate file
Moved country label logic to separate file to reduce length of App.js file.
