+++ src/App.js
@@ -28,6 +28,7 @@ import capitalData from "./assets/country_capitals.json"
 import { Button } from "semantic-ui-react"
 import CountrySearch from "./components/countrySearch.js"
 import oceaniaOutlines from "./components/oceaniaRegionOutlines.js"
+import countryLabels from "./components/countryLabels.js"
 
 // Arrays for label markers
 let countryMarkers = [];
@@ -89,6 +90,7 @@ class App extends Component {
     this.handleMapRefresh = this.handleMapRefresh.bind(this)
     this.handleDoubleClick = handleDoubleClick.bind(this)
     this.oceaniaOutlines = oceaniaOutlines.bind(this)
+    this.countryLabels = countryLabels.bind(this)
   }
 
   projection() {
@@ -376,53 +378,8 @@ class App extends Component {
                       )}
                     )}
                   </Geographies>
-
                   <Markers>{ this.oceaniaOutlines(countryMarkers) }</Markers>
-
-                  <Markers>
-                    {
-                      this.state.quiz ? this.state.quizGuesses.map((gss, i) => {
-                      if(gss){
-                        if(this.state.quizType === "name" || this.state.quizType === "flag" ) {
-                          var marker = countryMarkers.find(x => x.alpha3Code === this.state.quizAnswers[i]);
-                        } else if (this.state.quizType === "capital") {
-                          marker = capitalMarkers.find(x => x.alpha3Code === this.state.quizAnswers[i]);
-                        }
-                      }
-                      return gss&&(
-                        <Marker
-                          key={i}
-                          marker={marker}
-                          style={{
-                            default: { fill: "#FF5722" },
-                            hover: { fill: "#FFFFFF" },
-                            pressed: { fill: "#FF5722" },
-                          }}
-                        >
-                          {this.state.quizType === "capital" ? 
-                            (<circle
-                              cx={0}
-                              cy={0}
-                              r={2}
-                              className="dropFade"
-                              style={{
-                                stroke: "#FF5722",
-                                strokeWidth: 3,
-                                opacity: 0.9,
-                              }}
-                            />):null
-                          }
-                          <text
-                            textAnchor="middle"
-                            y={marker.markerOffset}
-                            className="mapLabel dropFade"
-                          >
-                            {marker.name}
-                          </text>
-                        </Marker>
-                      )}
-                    ):null}
-                  </Markers>
+                  <Markers>{ this.countryLabels(countryMarkers,capitalMarkers) }</Markers>
                 </ZoomableGroup>
               </ComposableMap>
               </div>
